[Feat] 포어그라운드/백그라운도/종료 앱 사이클에서의  FCM 딥링크 동작 구현

+++ TnT/Projects/Data/Sources/Network/Service/Trainer/TrainerRepositoryImpl.swift
@@ -50,7 +50,7 @@ public struct TrainerRepositoryImpl: TrainerRepository {
         return try await networkService.request(TrainerTargetType.getMonthlyLessonList(year: year, month: month), decodingType: GetMonthlyLessonListResDTO.self)
     }
     
-    public func getConnectedTraineeInfo(trainerId: Int, traineeId: Int) async throws -> GetConnectedTraineeInfoResponseDTO {
+    public func getConnectedTraineeInfo(trainerId: Int64, traineeId: Int64) async throws -> GetConnectedTraineeInfoResponseDTO {
         return try await networkService.request(TrainerTargetType.getConnectedTraineeInfo(trainerId: trainerId, traineeId: traineeId), decodingType: GetConnectedTraineeInfoResponseDTO.self)
     }
